Redesigned Tasks Page, Task Processing Function, and added images for better looking design
I have taken some design aspects from Coding2Go's To-Do Application. However, I have changed the task processing and implemented it in the way that I am familiar with. I still used some of the ways he did it, but for the rest, it was all my preferred method of implementation.

Will now look to deal with adding subtasks to the tasks, and then editing of the tasks (including deleting tasks). All these are the code mechanics that I will be looking into this week.