Refactored the CSS Design for the overlay and added a little JS functionality for the editing of temporarily unsaved assessment items
I made a few changes to the design of the overlay page. Added a delete icon which will be used when a user wishes to remove the assessment item before the save it temporarily.

Added a bit of JS functionality that would allow the user to save the changes they make to the temporary items or to cancel their changes if they desire.