Added the expanded module information overlay
This push includes the design for the overlay, that will display the module information details. The design includes the most important buttons and actions, including delete button, close information and edit information button.  The buttons don't yet have any functionality yet. This will be done at a later stage.

Will work on the design when editing the details of the module through the overlay