Added functionality to navigating modules page, adding modules, viewing module information
This push includes some much better functionality than previously. Functions are now independent from each other however. One function does not work in different scenarios, however, I should probably look into this near the end of the project

The functionalities include being able to open and close the overlays that display the input forms as well as the module information.

Arrays store the assessments and modules, however, they share a form of relationship as a result of having the same module codes