Editing and deleting of modules and assignments is now possible
When editing the module, the overlays pop up, allowing the user to now delete the module as a whole, with all of its related assessments.

Users can also select individual assessments to delete by once again opening the assessment input form through the add-module overlay.

Regarding the edits, the users can edit module information such as module information and lecturer information.

Assessments, however, cannot be edited. They can only be deleted or added. This may be inconvenient, but considering the time left till submission, this was the best I could do. Will plan on improving this implementation after the assessment.