Restructuring of folder; preparing for final versions
This push is mainly just moving things around so that they are a bit more organised. The main code remains outside of the folders, whilst the code that was a bit irrelevant has been moved to the other folders. The directory "Code due for further Development" includes a copy of the final files for submission, but I will work on the files in that directory when planning to make further changes and improvements, without affecting the one that is currently working.

Then if there are any features that I deem are ready to be added to the final project, then I will simply apply these changes to the final project from the experimental of further developed code